flask-mysql-zappa-lambda/requirements.txt
argcomplete==3.4.0
blinker==1.8.2
boto3==1.34.132
botocore==1.34.132
certifi==2024.6.2
cfn-flip==1.3.0
charset-normalizer==3.3.2
click==8.1.7
durationpy==0.6
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
gitdb==4.0.11
GitPython==3.1.41
greenlet==3.0.3
hjson==3.1.0
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
jmespath==1.0.1
kappa==0.6.0
MarkupSafe==2.1.5
mysqlclient==2.2.4
placebo==0.9.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-slugify==8.0.4
PyYAML==6.0.1
requests==2.32.3
s3transfer==0.10.2
six==1.16.0
smmap==5.0.1
SQLAlchemy==2.0.31
text-unidecode==1.3
toml==0.10.2
tqdm==4.66.4
troposphere==4.8.0
typing_extensions==4.12.2
urllib3==2.2.2
Werkzeug==3.0.3
zappa==0.59.0
